modified:   a.out 	deleted:    background-stuff/GUI_interface/GUI_interface.pde 	deleted:    background-stuff/GUI_interface/boot_screen.pde 	deleted:    background-stuff/GUI_interface/data/gui/GUI_interface/saves/auto.json 	deleted:    background-stuff/GUI_interface/diag_screen.pde 	deleted:    background-stuff/GUI_interface/pdf_print.pde 	deleted:    background-stuff/GUI_interface/preven_screen.pde 	deleted:    background-stuff/Logic_Check_Order.txt deleted:    background-stuff/chatgippitydiag.cpp 	deleted:    background-stuff/diag.h 	deleted:    background-stuff/expected-diag.cpp 	deleted:    background-stuff/fixing-prevent.cpp 	deleted:    background-stuff/fixing_diag.cpp 	deleted:    background-stuff/headers.h 	deleted:    background-stuff/logic-all.txt 	deleted:    background-stuff/prev_after.txt 	deleted:    background-stuff/prev_all.txt 	deleted:    background-stuff/prev_before.txt 	deleted:    background-stuff/prev_during.txt 	deleted:    background-stuff/prevent.h deleted:    background-stuff/testing.cpp 	modified:   diagnostic.cpp 	modified:   main.cpp 	new file:   output.txt 	modified:   univ-function.cpp 	new file:   vehicle-directory/1.txt
