--- background-stuff/GUI_interface/GUI_interface.pde
@@ -1,18 +1,0 @@
-/*import com.krab.lazy.*;
-import processing.pdf.*;
-
-LazyGui gui;
-
-void setup(){
-    size(800,800,P2D);
-    gui = new LazyGui(this);
-}
-
-void draw(){
-    background(gui.colorPicker("background").hex);
-}
-*/
-void setup() {
-  size(100, 100);
-  parseFile();
-}
