--- background-stuff/prev_during.txt
@@ -1,22 +1,0 @@
-    While vehicle's running/driving:
-        if the engine suddently shuts off:
-            you either stalled (shift better if you're in a manual or replace transmission if in a automatic) or your engine needs to be replaced.
-        if the car doesn't stay in a straight line
-            check tie rods
-                if they're bent, you can attempt to straighten them to get home (not recommended), but replace ASAP.
-                if not, get an alignment
-        if the car is incredibly "bouncy" over bumps:
-            Your suspension is too old, replace suspension
-        if the car pulls to the left or right while braking:
-            flush brake fluid, if that doesn't fix it
-            check for (and replace) seized brake cylinders, if none are siezed check and replace brake fluid pump
-        if the car doesn't stop when braking
-            check and replace brake pads
-            flush brake fluid, if that doesn't fix it
-            check for (and replace) seized brake cylinders, if none are siezed check and replace brake fluid pump
-        if the car shifts hard or doesn't change gear:
-            transmission or clutch is going, recommend seeing a mechanic.
-        if the car vibrates heavily when stopped
-            the motor mounts are going or mis-sized, see mechanic to replace.
-        if vehicle doesn't move when pushing on the gas and in gear
-            either the clutch is out or the drivetrain is disconnected, tow to mechanic.
