+++ univ-function.cpp
@@ -22,3 +22,25 @@ string get_vehicle_number() {
 
   return VehicleNum;
 }
+
+int UserInput(int valid) {
+  int userIn;
+  cin >> userIn;
+
+  while (userIn < 0 || userIn > valid) {
+    cout << "Invalid input! Please select from 1 to " << valid << endl;
+    cin >> userIn;
+  }
+
+  ofstream outFile("output.txt", ios::trunc); // Open file in trunc mode
+
+  if (!outFile) {
+      cout << "Error opening file!" << endl;
+  }
+
+  int number = valid; // Get valid input from UserInput()
+  outFile << number << endl; // Store it in file
+  outFile.close();
+
+  return userIn;
+}
