+++ vehicle-directory/1.txt
@@ -1,0 +1,7 @@
+Check on vehicle 1 performed on 2025-3-2
+
+Be aware: battery problems.
+Issue unknown, third-party inspection required.
+Diagnostic report for 1 vehicle:
+Issue reported: Engine Accessories
+Diagnostic report for 1 vehicle:
